+++ src/main/java/efd/contribuicoes/blocoD/n1/n2/n3/RegD100.java
@@ -41,7 +41,7 @@ public class RegD100 implements Serializable {
     private String sub;
 
     @Getter @Setter @Inclui
-    private Integer numDoc;
+    private Long numDoc;
 
     @Getter @Setter @Inclui
     private String chvCte;
@@ -94,7 +94,7 @@ public class RegD100 implements Serializable {
     @Getter @Setter @Filho
     private List<RegD111> regD111;
 
-    public RegD100(String indOper, String indEmit, String codPart, String codMod, Integer codSit, String ser, String sub, Integer numDoc, String chvCte, LocalDate dtDoc, LocalDate dtAP, Integer tpCte, String chvCteRef, BigDecimal vlDoc, BigDecimal vlDesc, String indFrt, BigDecimal vlServ, BigDecimal vlBcIcms, BigDecimal vlIcms, BigDecimal vlNt, String coInf, String codCta) {
+    public RegD100(String indOper, String indEmit, String codPart, String codMod, Integer codSit, String ser, String sub, Long numDoc, String chvCte, LocalDate dtDoc, LocalDate dtAP, Integer tpCte, String chvCteRef, BigDecimal vlDoc, BigDecimal vlDesc, String indFrt, BigDecimal vlServ, BigDecimal vlBcIcms, BigDecimal vlIcms, BigDecimal vlNt, String coInf, String codCta) {
         this.indOper = indOper;
         this.indEmit = indEmit;
         this.codPart = codPart;
