+++ src/main/java/efd/icmsipi/blocoD/n1/n2/RegD300.java
@@ -29,10 +29,10 @@ public class RegD300 implements Serializable {
     private Integer sub;
 
     @Getter @Setter @Inclui
-    private Integer numDocIni;
+    private Long numDocIni;
 
     @Getter @Setter @Inclui
-    private Integer numDocFin;
+    private Long numDocFin;
 
     @Getter @Setter @Inclui(zerosEsquerda = 3)
     private Integer cstIcms;
@@ -82,7 +82,7 @@ public class RegD300 implements Serializable {
     @Getter @Setter @Filho
     private List<RegD310> regD310;
 
-    public RegD300(String codMod, String ser, Integer sub, Integer numDocIni, Integer numDocFin, Integer cstIcms, Integer cfop, BigDecimal aliqIcms, LocalDate dtDoc, BigDecimal vlOpr, BigDecimal vlDesc, BigDecimal vlServ, BigDecimal vlSeg, BigDecimal vlOutDesp, BigDecimal vlbcIcms, BigDecimal vlIcms, BigDecimal vlRedBc, String codObs, String codCta) {
+    public RegD300(String codMod, String ser, Integer sub, Long numDocIni, Long numDocFin, Integer cstIcms, Integer cfop, BigDecimal aliqIcms, LocalDate dtDoc, BigDecimal vlOpr, BigDecimal vlDesc, BigDecimal vlServ, BigDecimal vlSeg, BigDecimal vlOutDesp, BigDecimal vlbcIcms, BigDecimal vlIcms, BigDecimal vlRedBc, String codObs, String codCta) {
         this.codMod = codMod;
         this.ser = ser;
         this.sub = sub;
