+++ src/main/java/efd/icmsipi/blocoE/n1/n2/n3/n4/n5/RegE113.java
@@ -28,7 +28,7 @@ public class RegE113 implements Serializable {
     private Integer sub;
 
     @Getter @Setter @Inclui
-    private Integer numDoc;
+    private Long numDoc;
 
     @Getter @Setter @Inclui
     private LocalDate dtDoc;
@@ -42,7 +42,7 @@ public class RegE113 implements Serializable {
     @Getter @Setter @Inclui
     private String chvDoce;
 
-    public RegE113(String codPart, String codMod, String ser, Integer sub, Integer numDoc, LocalDate dtDoc, String codItem, BigDecimal vlAjItem, String chvDoce) {
+    public RegE113(String codPart, String codMod, String ser, Integer sub, Long numDoc, LocalDate dtDoc, String codItem, BigDecimal vlAjItem, String chvDoce) {
         this.codPart = codPart;
         this.codMod = codMod;
         this.ser = ser;
